Refactor raft/grpc TLS configuration & update image tag
This commit moves all TLS related raft & grpc code into its own file. It also
updates the kustomize image tag to the next version.
