+++ deploy/kustomize/kustomization.yaml
@@ -4,7 +4,7 @@ namespace: arrebato
 
 images:
   - name: ghcr.io/davidsbond/arrebato
-    newTag: v0.1.0-alpha1
+    newTag: v0.1.0-alpha2
 
 resources:
   - namespace.yaml
